backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==5.5.2
qdrant-client==1.8.0
google-generativeai==0.4.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3